+++ b4n-tasks/commands/list_themes.rs
@@ -1,4 +1,4 @@
-use b4n_config::Config;
+use b4n_config::{Config, Persistable, themes::Theme};
 use std::path::PathBuf;
 use tokio::fs;
 
@@ -8,24 +8,33 @@ use crate::commands::CommandResult;
 pub struct ListThemesCommand;
 
 impl ListThemesCommand {
-    /// Gets all files from the themes directory.
+    /// Gets all files from the themes directory.\
+    /// **Note** that it includes `default` theme if not present.
     pub async fn execute(&self) -> Option<CommandResult> {
-        if let Ok(list) = get_themes().await {
-            Some(CommandResult::ThemesList(list))
-        } else {
-            None
+        let default = Theme::default_path();
+        let path = Config::themes_dir();
+        let mut list = list_themes(&path).await.unwrap_or_default();
+        if !list.contains(&default) {
+            list.push(default);
         }
+
+        Some(CommandResult::ThemesList(list))
     }
 }
 
-async fn get_themes() -> Result<Vec<PathBuf>, std::io::Error> {
+async fn list_themes(path: &PathBuf) -> Result<Vec<PathBuf>, std::io::Error> {
     let mut result = Vec::new();
-    let path = Config::themes_dir();
     let mut dir = fs::read_dir(path).await?;
+
     while let Some(entry) = dir.next_entry().await? {
         let file_type = entry.file_type().await?;
         if file_type.is_file() {
-            result.push(entry.path());
+            let path = entry.path();
+            if let Some(extension) = path.extension()
+                && (extension.eq_ignore_ascii_case("yaml") || extension.eq_ignore_ascii_case("yml"))
+            {
+                result.push(path);
+            }
         }
     }